Finishes Sign up and Sign in stories
Many changes have been made. I made it so that users input their first
and last names when they sign up. I also added a users controller and a
simple ‘show’ view, just so that there is a root (which is needed for
my test). I also completed the necessary steps for my cucumber tests.
The two tests that go with my features are 100% green, and I have 100%
LOC coverage.