Creating and signing for events
Admin now have the ability to create new volunteering events, and users
have the ability to sign up for them. When a volunteer signs up for an
event, this can be seen by the admin through the admin dashboard.
Approved volunteers will see a list of all upcoming volunteer events on
their home page, and if they sign up, the listing will turn blue on the
home page.