Updated view
Updated view of the volunteer form. TO-DO List for Kevin: Create a
migration for the volunteer_application model to include all of the
necessary fields, break the form up into multiple parts on different
pages so that the params list isn’t too long, and update the cucumber
test to reflect the changes.